+++ out/getHtml.js
@@ -395,6 +395,10 @@ function createEditorHtml(webview, context, initialVars) {
 													<div class="field mar-left">
 														<label>
 															<span>EscapeChar</span>
+															<span class="tooltip is-tooltip-multiline"
+																data-tooltip="The character used to escape the QuoteChar inside field values">
+																<i class="fas fa-question-circle"></i>
+															</span>
 														</label>
 														<input id="escape-char-string" class="input" type="text" oninput="setEscapeCharString()">
 													</div>
@@ -469,6 +473,10 @@ function createEditorHtml(webview, context, initialVars) {
 												<div class="field mar-left">
 													<label>
 														<span>EscapeChar</span>
+														<span class="tooltip is-tooltip-multiline"
+															data-tooltip="The character used to escape the QuoteChar inside field values">
+															<i class="fas fa-question-circle"></i>
+														</span>
 													</label>
 													<input id="escape-char-string-write" class="input" type="text" oninput="setEscapeCharStringWrite()">
 												</div>
@@ -522,7 +530,7 @@ function createEditorHtml(webview, context, initialVars) {
 							<i class="fas fa-chevron-down left-panel-toggle-icon-collapse"></i>
 						</div>
 					</div>
-
+					
 					<button class="button is-outlined" onclick="addRow()">
 						<span class="icon is-small">
 							<i class="fas fa-plus"></i>
@@ -609,7 +617,6 @@ function createEditorHtml(webview, context, initialVars) {
 					<progress id="received-csv-prog-bar" class="progress is-info" value="50" max="100"></progress>
 				</div>
 			</div>
-
 
 
 			<!-- main editor/grid area -->